Automation Testing code based on Selenium
The folder contains code to test the functioning of the dashboard and obtain bugs in case there are issues. The code tests for country buttons and maps, as well as working of the header sections (Timeline, Demographic, Impact, Events). For visual testing, a result image with the regions which mismatch with the baseline images is obtained